requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
aiofiles>=23.2.0
python-dotenv>=1.0.0
